Move menu partials to components
This way the code can be more consistent.

IMHO it's pretty safe to refactor admin, moderation, management,
valuation and officing views without harming institutions who might have
customized this views, simply because these sections are usually not
customized at all since institutions prioritize the public views.

Note the officing menu is slightly different since it depends on a
variable set in the controller.
