+++ app/components/moderation/menu_component.html.erb
