+++ app/components/moderation/menu_component.rb
@@ -1,0 +1,3 @@
+class Moderation::MenuComponent < ApplicationComponent
+  include LinkListHelper
+end
