+++ app/components/valuation/menu_component.rb
@@ -1,0 +1,3 @@
+class Valuation::MenuComponent < ApplicationComponent
+  include LinkListHelper
+end
